requirements.txt
absl-py==2.3.1
anytree==2.13.0
asttokens==3.0.0
attrs==25.3.0
avp_stream==1.0
certifi==2025.8.3
charset-normalizer==3.4.3
cloudpickle==3.1.1
cmeel==0.57.3
cmeel-assimp==5.4.3.1
cmeel-boost==1.87.0.1
cmeel-console-bridge==1.0.2.3
cmeel-octomap==1.10.0
cmeel-qhull==8.0.2.1
cmeel-tinyxml2==10.0.0
cmeel-urdfdom==4.0.1
cmeel-zlib==1.3.1
coal==3.0.1
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
dex_retargeting==0.5.0
dm_control==1.0.31
dm-env==1.6
dm-tree==0.1.9
docstring_parser==0.17.0
eigenpy==3.10.3
etils==1.13.0
executing==2.2.1
Farama-Notifications==0.0.4
filelock==3.13.1
fonttools==4.59.1
fsspec==2025.7.0
glfw==2.9.0
grpcio==1.74.0
grpcio-tools==1.74.0
gymnasium==1.2.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_resources==6.5.2
ipdb==0.13.13
ipython==9.6.0
ipython_pygments_lexers==1.1.1
isaacgym==1.0rc4
jedi==0.19.2
Jinja2==3.1.4
kiwisolver==1.4.9
labmaze==1.0.6
llvmlite==0.45.1
lxml==6.0.0
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
mujoco==3.3.3
mujoco-python-viewer==0.1.4
networkx==3.3
nlopt==2.9.1
numba==0.62.1
numpy==2.2.6
nvidia-cublas-cu12==12.4.2.65
nvidia-cuda-cupti-cu12==12.4.99
nvidia-cuda-nvrtc-cu12==12.4.99
nvidia-cuda-runtime-cu12==12.4.99
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.0.44
nvidia-curand-cu12==10.3.5.119
nvidia-cusolver-cu12==11.6.0.99
nvidia-cusparse-cu12==12.3.0.142
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.4.99
opencv-python==4.12.0.88
packaging==25.0
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
pin==3.7.0
pip==25.1
prompt_toolkit==3.0.52
protobuf==6.32.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
PyOpenGL==3.1.10
PyOpenGL-accelerate==3.1.10
pyparsing==3.2.3
pyperclip==1.11.0
python-dateutil==2.9.0.post0
pytransform3d==3.14.2
PyYAML==6.0.2
requests==2.32.5
rich==14.1.0
sapien==3.0.1
scipy==1.16.1
setuptools==78.1.1
shtab==1.7.2
six==1.17.0
stack-data==0.6.3
sympy==1.13.3
tqdm==4.67.1
traitlets==5.14.3
transforms3d==0.4.2
triton==3.0.0
typeguard==4.4.4
typing_extensions==4.14.1
tyro==0.9.32
urllib3==2.5.0
wcwidth==0.2.14
wheel==0.45.1
wrapt==1.17.3
zipp==3.23.0